+++ tests/test_operations.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 from typing import Any, Dict, Type
 
 from app.exceptions import ValidationError
-from app.operations import (
+from app.operations.operations import (
     Operation,
     Addition,
     Subtraction,
